backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
aiosqlite==0.20.0
aiofiles==24.1.0
openpyxl==3.1.5
